Move functions from protractor to `math/`
