+++ packages/perseus/src/widgets/interactive-graphs/graphs/utils.ts
@@ -1,10 +1,6 @@
 import type {CollinearTuple} from "../../../perseus-types";
 import type {Interval, vec} from "mafs";
 
-export function calculateAngleInDegrees([x, y]: vec.Vector2) {
-    return (Math.atan2(y, x) * 180) / Math.PI;
-}
-
 /**
  * Given a ray and a rectangular box, find the point where the ray intersects
  * the edge of the box. Assumes the `initialPoint` is inside the box.
