+++ packages/perseus/src/widgets/interactive-graphs/math/index.ts
@@ -3,4 +3,5 @@ export {snap} from "./snap";
 export {inset, clampToBox} from "./box";
 export {X, Y} from "./coordinates";
 export {MIN, MAX, size} from "./interval";
-export {findAngle, polar} from "./angles";
+export {lerp} from "./interpolation";
+export {calculateAngleInDegrees, convertDegreesToRadians, findAngle, polar} from "./angles";
